requirements.txt
matplotlib==3.8.0
numpy==1.23.5
pandas==2.2.2
python-dotenv==1.0.1
ratelimit==2.2.1
Requests==2.31.0
tqdm==4.64.1
python==3.10